Make effect `amplifier` type `VarInt` and update effect color calc

+++ examples/potions.rs
@@ -151,14 +151,14 @@ pub fn add_potion_effect(
     }
 }
 
-fn adjust_modifier_amount(amplifier: u8, amount: f64) -> f64 {
+fn adjust_modifier_amount(amplifier: i32, amount: f64) -> f64 {
     amount * f64::from(amplifier + 1)
 }
 
 fn apply_potion_attribute(
     attributes: &mut Mut<EntityAttributes>,
     health: &mut Option<Mut<Health>>,
-    amplifier: u8,
+    amplifier: i32,
     attr: AttributeModifier,
     name: &str,
 ) {
@@ -222,7 +222,7 @@ pub fn handle_status_effect_added(
                     // need to keep track of the previous absorption value and subtract that from
                     // the new value (because you can take damage while having absorption)
                     if let Some(mut absorption) = absorption {
-                        absorption.0 += f32::from(effect.amplifier() + 1) * 4.0;
+                        absorption.0 += (effect.amplifier() + 1) as f32 * 4.0;
                     }
                 }
                 StatusEffect::InstantHealth => {
@@ -274,7 +274,7 @@ pub fn handle_status_effect_removed(
             match effect.status_effect() {
                 StatusEffect::Absorption => {
                     if let Some(mut absorption) = absorption {
-                        absorption.0 -= f32::from(effect.amplifier() + 1) * 4.0;
+                        absorption.0 -= (effect.amplifier() + 1) as f32 * 4.0;
                     }
                 }
                 StatusEffect::Glowing => {
@@ -305,7 +305,7 @@ pub fn handle_status_effect_update(
         for effect in status.get_current_effects() {
             match effect.status_effect() {
                 StatusEffect::Regeneration => {
-                    let i = 50 >> u32::from(effect.amplifier().min(31));
+                    let i = 50 >> effect.amplifier().min(31);
 
                     if i == 0 || effect.active_ticks() % i == 0 {
                         if let Some(ref mut health) = health {
@@ -318,7 +318,7 @@ pub fn handle_status_effect_update(
                     }
                 }
                 StatusEffect::Poison => {
-                    let i = 25 >> u32::from(effect.amplifier().min(31));
+                    let i = 25 >> effect.amplifier().min(31);
 
                     if i == 0 || effect.active_ticks() % i == 0 {
                         if let Some(ref mut health) = health {
@@ -327,7 +327,7 @@ pub fn handle_status_effect_update(
                     }
                 }
                 StatusEffect::Wither => {
-                    let i = 40 >> u32::from(effect.amplifier().min(31));
+                    let i = 40 >> effect.amplifier().min(31);
 
                     if i == 0 || effect.active_ticks() % i == 0 {
                         if let Some(ref mut health) = health {
